Create comment object
