+++ src/styles/main.css
@@ -79,7 +79,7 @@ footer p {
 }
 
 .comment .book-cover {
-  width: 60%;
+  width: 100%;
   height: 60vh;
 }
 
@@ -95,7 +95,7 @@ form {
 }
 
 input {
-  height: 1rem;
+  height: 2rem;
 }
 
 .close {
